requirements.txt
fastapi
uvicorn[standard]
python-multipart
jinja2
aiofiles
pillow
opencv-python
python-dotenv
